fix: only decode proper CIDs as CIDs
Basically, I spent hours digging through randoms serde stuff and this is
as close as I can get to idiomatic upstream.

The main difference is: upstream doesn't really support "deserialize
any" for newtypes. So we have to hack it a bit and just assume that we
only have one newtype (which, luckily, is correct).